Move container fixture to pytest conftest file
